d68k:  Ready get_aTrap_name() for additional logic
Prepare for the next patch, which adds logic to get_aTrap_name().  Aside
from making it not inline, store the result temporarily so we can act on
it.  Also, call get_trap_name() directly in decode_MOVE(), circumventing
the new logic for trap name lookup failure, which we disregard for MOVE.