icon2skif:  Make wayward pixels black, not empty
Before the advent of System 7, when the Finder only displayed icons with
one bit per pixel (with one very colorful exception, thanks to a handful
of trap patches and a whole lot of cleverness), the process of rendering
an icon consisted of two blitting operations: applying a mask (using Bic
mode) and then the icon's face, with Xor.  This meant that pixels set in
the icon face could be used not only to paint on top of the destination,
but to interact with it -- turning black pixels white as well as turning
white ones black.  In practice, though, icons mostly appeared on a white
background, so an area within an icon that inverted the background would
just show up as black.  However, the all-white background afforded a new
opportunity: creating icons that, when selected (thus inverting the bits
in the mask), displayed as another icon, whose relationship to the first
is entirely arbitrary -- the purest example of which might be Dungeon of
Doom's main icon, which initially appears as a pedestal-mounted orb, but
while selected displays as a sword and shield.  Another notable instance
is StuntCopter's icon, depicting a stunt jumper either dangling from the
chopper or having just landed on the ground.  There are others, besides.

We label pixels which are set in the face but clear in the mask with the
term "wayward".  Previously, we were treating wayward pixels the same as
"outside" pixels (which are clear in both face and mask), rendering them
into fully transparent ARGB pixels.  But this yields poor results if the
icon contains wayward pixels.  The most absurd case is Shufflepuck Café,
whose icon face is drawn entirely with wayward pixels, thus rendering as
completely white when those pixels are masked out (as we were doing, and
as they are in Mac OS 9).  To cope with this, render wayward pixels into
opaque black ARGB pixels instead.  In effect, we're constructing an icon
"base" as (face OR mask), which we use in place of the actual icon mask.