vlib:  Don't unshare non-pairs in first_mutable()
Not only might there be strings that don't need to be unshared, but also
unique types (e.g. auto fd, thread, channel) for which it's meaningless.
Unsharing one results in two objects pointing to the same resource, each
of which is set to destroy it -- yielding Undefined Behavior (or worse).