varyx-posix:  Nix duplication of threads, auto fds
These objects represent actual, external resources that can't be copied.

(You can trivially mutate fd 42 into fd 54, because it's just an integer
value with a semantic tag.  But auto fd 42 is a load-bearing object that
will close(42) when it's destroyed.  It's wrong for that to occur twice,
and there shouldn't be a way to unshare the auto fd in the first place.)