stack-{space,chain}:  Fix _relix_stack_space()
So, we stopped setting the stack limit field in the user parameter block
because it's a per-process structure, and storing per-thread information
in it breaks in the presence of multithreading.  But that leaves us with
a purported stack limit of zero, so a stack space check always succeeds.

(The check can't even fail theoretically due to the SP approaching zero,
because you would have crashed long before then, from a clobbered heap.)

Instead, call a kernel utility routine to get the available stack space.

This fixes the regression from c6e472b555de603d5a86fba104d9b812a1b85fbd.